Refactor test structure: remove old test set management tests and add new app management tests
- Deleted `manage.spec.ts` and `manage.ts` files for outdated test set management.
- Introduced new `app-management.spec.ts` and `app-management.ts` files for app management tests.
- Added comprehensive tests for app deletion and renaming functionalities.
- Created new test set management tests in `testset-management.spec.ts` and `testset-management.ts` to cover various scenarios including creation, editing, and deletion of test sets.
