+++ web/oss/tests/playwright/5-testsset-manage.ts
@@ -1,3 +1,3 @@
-import testsetManageTests from "./acceptance/testsset/manage"
+import testsetManageTests from "./acceptance/testsset/testset-management"
 
 export default testsetManageTests
