+++ web/oss/tests/playwright/acceptance/app/app-management.spec.ts
@@ -1,0 +1,3 @@
+import tests, {test} from "./app-management"
+
+test.describe(`App Management`, tests)
