--- web/oss/tests/playwright/acceptance/app/manage.spec.ts
@@ -1,3 +1,0 @@
-import tests, {test} from "./manage"
-
-test.describe(`App Management`, tests)
