+++ web/oss/tests/playwright/acceptance/testsset/testset-management.spec.ts
@@ -1,4 +1,4 @@
-import tests from "./manage"
+import tests from "./testset-management"
 import {test} from "@agenta/web-tests/tests/fixtures/base.fixture"
 
 test.describe(`Test Sets`, tests)
